Seperate StepCard and StepHeader of Main.tsx
