+++ client/web/src/components/GrowBox.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const GrowBox = React.memo(
+  () => <div className="grow flex" />,
+  () => true,
+);
+
+export default GrowBox;
