+++ client/web/src/types/main.ts
@@ -1,0 +1,11 @@
+export type StepCardType = {
+  id: number;
+  name: string;
+  subTitle: string;
+  rep: number;
+  rest: number;
+  set: number;
+  imageUrl: string;
+  detail: string;
+  status: string;
+};
